+++ supplier_api_tester/supplier_api_tester/tests/__init__.py
@@ -25,7 +25,6 @@ RESERVATION_TEST = (
     reservation.test_incorrect_date_format,
     reservation.test_past_date,
     reservation.test_not_allowed_method,
-    # TODO check for response w/error 2006?
 )
 
 BOOKING_TEST = (
@@ -35,7 +34,7 @@ BOOKING_TEST = (
     booking.test_not_allowed_method,
     booking.test_booking_incorrect_reservation_id,
     booking.test_booking,
-    booking.test_cancellation
+    booking.test_cancellation,
 )
 
 PRODUCT_CATALOG = (