+++ supplier_api_tester/supplier_api_tester/utils/adapters.py
@@ -59,25 +59,26 @@ def parse_availability_variants(raw_response: Response, response: Dict) -> List[
     available_variants: List[DailyVariants] = []
     try:
         for day, day_availability in response.items():
-            for timeslot, timeslot_availability in day_availability.items():
-                available_variants.append(
-                    dacite.from_dict(
-                        data_class=DailyVariants,
-                        data={
-                            'date': day,
-                            'timeslot': timeslot,
-                            'available_tickets': timeslot_availability.get('available_tickets'),
-                            'variants': [
-                                {**v}
-                                for v in timeslot_availability.get('variants', [])
-                            ]
-                        },
-                        config=dacite.Config(
-                            type_hooks={date: date.fromisoformat},
-                            strict=True,
-                        )
+            # extract timeslot component from the day value: YYYY-MM-DDTHH:MM
+            date_time_value = datetime.fromisoformat(day)
+            available_variants.append(
+                dacite.from_dict(
+                    data_class=DailyVariants,
+                    data={
+                        'date': str(date_time_value.date()),
+                        'timeslot': f'{date_time_value.time().hour}:{date_time_value.time().minute}',
+                        'available_tickets': day_availability.get('available_tickets', 0),
+                        'variants': [
+                            {**v}
+                            for v in day_availability.get('variants', [])
+                        ]
+                    },
+                    config=dacite.Config(
+                        type_hooks={date: date.fromisoformat},
+                        strict=True,
                     )
                 )
+            )
     except (
         dacite.exceptions.WrongTypeError,
         dacite.exceptions.MissingValueError,
@@ -87,7 +88,6 @@ def parse_availability_variants(raw_response: Response, response: Dict) -> List[
             message=f'Incorrect JSON format in response from the /availability endpoint: {format_error_message(e)}',
             response=raw_response,
         )
-
     return available_variants
 
 
@@ -119,7 +119,7 @@ def get_products(raw_response: Response, response: Dict) -> List[Product]:
 
 
 def get_reservation(raw_response: Response, response) -> Reservation:
-    '''Getting and testing response from the /reservation endpoint'''
+    """Getting and testing response from the /reservation endpoint"""
     if type(response) is not dict:
         raise FailedTest(
             message='The response should be a JSON Object',
@@ -146,7 +146,7 @@ def get_reservation(raw_response: Response, response) -> Reservation:
 
 
 def get_booking(raw_response: Response, response) -> Booking:
-    '''Getting and testing response from the /booking endpoint'''
+    """Getting and testing response from the /booking endpoint"""
     if type(response) is not dict:
         raise FailedTest(
             message='The response should be a JSON Object',
@@ -197,7 +197,12 @@ def get_booking(raw_response: Response, response) -> Booking:
 
 
 def get_api_error(raw_response: Response, response) -> ApiError:
-    '''Unpacking 400 error JSON structure'''
+    """Unpacking 400 error JSON structure"""
+    if raw_response.ok:
+        raise FailedTest(
+            message=f'Expected HTTP 400 but got HTTP {response.get("status_code")} instead.',
+            response=raw_response,
+        )
     if type(response) is not dict:
         raise FailedTest(
             message='400 error response should be a JSON Object',