+++ supplier_server_mock/supplier_server/constants.py
@@ -6,11 +6,9 @@ PRODUCTS = [
         'is_refundable': True,
         'cutoff_time': 24,
         'max_tickets_per_order': 10,
-        'required_visitor_data': [],
-        'required_order_data': [],
         # but these fields are not part of spec...
         '_ticket_content_type': 'CODE128',
-        '_cancelled_bookings': [],
+        '_cancelled_bookings': {},
         '_timeslot_available_tickets_as_sum': True  # available tickets is the sum among all variants
     },
     {
@@ -24,7 +22,7 @@ PRODUCTS = [
         'required_visitor_data': ['FULL_NAME', 'PHONE'],
         'required_order_data': ['PICKUP_LOCATION', 'PASSPORT_ID'],
         '_ticket_content_type': 'CODE128',
-        '_cancelled_bookings': [],
+        '_cancelled_bookings': {},
         '_timeslot_available_tickets_as_sum': False
     },
     {
@@ -35,10 +33,9 @@ PRODUCTS = [
         'is_refundable': True,
         'cutoff_time': 0,
         'max_tickets_per_order': 25,
-        'required_visitor_data': [],
         'required_order_data': ['PICKUP_LOCATION', 'PASSPORT_ID', 'FLIGHT_NUMBER'],
         '_ticket_content_type': 'CODE128',
-        '_cancelled_bookings': [],
+        '_cancelled_bookings': {},
         '_timeslot_available_tickets_as_sum': False
     },
     {
@@ -49,9 +46,8 @@ PRODUCTS = [
         'is_refundable': True,
         'cutoff_time': 10,
         'required_visitor_data': ['EMAIL', 'DATE_OF_BIRTH'],
-        'required_order_data': [],
         '_ticket_content_type': 'PDF',
-        '_cancelled_bookings': [],
+        '_cancelled_bookings': {},
         '_timeslot_available_tickets_as_sum': False
     },
     {
@@ -61,10 +57,9 @@ PRODUCTS = [
         'is_refundable': False,
         'cutoff_time': 0,
         'max_tickets_per_order': 5,
-        'required_visitor_data': [],
         'required_order_data': ['NATIONALITY'],
         '_ticket_content_type': 'CODE128',
-        '_cancelled_bookings': [],
+        '_cancelled_bookings': {},
         '_timeslot_available_tickets_as_sum': False
     },
 ]