+++ supplier_api_tester/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.0.0
+
+- add support for API v2.0.0
+- drop support for API v1.x
+- drop support for the CLI argument `--timeslots` (`-t`) in v2.0 
+
 ## 1.6.5
 
 - Using time-based reference_id for booking